+++ client/src/components/Form/Reservation/PickDate.js
@@ -1,39 +1,36 @@
 import { DatePicker, TimePicker } from 'antd';
 import moment from 'moment';
-import PropTypes from 'prop-types';
 import React from 'react';
+import ThemeContext from './Context';
 
-const DatePick = ({ repeatValue, handleChange, timeHandle, dateROnChange }) => {
+const DatePick = () => {
   const disabledDate = (current) => current && current < moment().endOf('day');
   // Can not select days before today and today
 
   return (
-    <div>
-      {repeatValue !== 'once' ? (
-        <DatePicker.RangePicker
-          onChange={dateROnChange}
-          disabledDate={disabledDate}
-        />
-      ) : (
-        <DatePicker
-          format="YYYY-MM-DD"
-          disabledDate={disabledDate}
-          onChange={handleChange}
-        />
+    <ThemeContext.Consumer>
+      {({ repeat, dateOnChange, timeOnChange, dateROnChange }) => (
+        <div>
+          {repeat !== 'once' ? (
+            <DatePicker.RangePicker
+              onChange={dateROnChange}
+              disabledDate={disabledDate}
+            />
+          ) : (
+            <DatePicker
+              format="YYYY-MM-DD"
+              disabledDate={disabledDate}
+              onChange={dateOnChange}
+            />
+          )}
+          <br />
+          <br />
+
+          <TimePicker.RangePicker onChange={timeOnChange} />
+        </div>
       )}
-      <br />
-      <br />
-
-      <TimePicker.RangePicker onChange={timeHandle} />
-    </div>
+    </ThemeContext.Consumer>
   );
-};
-
-DatePick.propTypes = {
-  repeatValue: PropTypes.string.isRequired,
-  handleChange: PropTypes.func.isRequired,
-  timeHandle: PropTypes.func.isRequired,
-  dateROnChange: PropTypes.func.isRequired,
 };
 
 export default DatePick;
